Fix integer duration

+++ scripts/waypoints_test.py
@@ -23,7 +23,7 @@ def parse_args():
 
 
 def move_forward(distance, left_speed, right_speed, forward_rate=0.415):
-    duration = round(distance/forward_rate * 10)
+    duration = int(round(distance/forward_rate * 10))
 
     for _ in range(duration):
 
@@ -38,7 +38,7 @@ def move_forward(distance, left_speed, right_speed, forward_rate=0.415):
 
 
 def right_turn(turn, left_speed, right_speed, turn_rate=3.02):
-    duration = round(turn/turn_rate * 10)
+    duration = int(round(turn/turn_rate * 10))
 
     for _ in range(duration):
 
@@ -54,7 +54,7 @@ def right_turn(turn, left_speed, right_speed, turn_rate=3.02):
 
 def left_turn(turn, left_speed, right_speed, turn_rate=1.87):
 
-    duration = round(turn/turn_rate * 10)
+    duration = int(round(turn/turn_rate * 10))
 
     for _ in range(duration):
